保存平台商品SPU

+++ toucan-shopping-modules/product/toucan-shopping-product-entity-vo/src/main/java/com/toucan/shopping/modules/product/entity/ProductSpu.java
@@ -27,7 +27,7 @@ public class ProductSpu {
 
     @JsonFormat(shape=JsonFormat.Shape.STRING)
     @JSONField(serializeUsing = ToStringSerializer.class)
-    private Integer categoryId; //所属类别
+    private Long categoryId; //所属类别
 
     private String uuid; //SPU UUID
